pyproject.toml
[project.urls]
Homepage = "https://github.com/treeolivetech/pkg-tawala"
Documentation = "https://github.com/treeolivetech/pkg-tawala#readme"
Repository = "https://github.com/treeolivetech/pkg-tawala"
Issues = "https://github.com/treeolivetech/pkg-tawala/issues"
# Changelog = "https://github.com/treeolivetech/pkg-tawala/blob/main/CHANGELOG.md"

[project]
name = "Tawala"
version = "1.3.0"
description = "Build and deploy Django apps with confidence."
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "Treeolive Technologies", email = "hi@treeolive.tech" }]
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Framework :: Django :: 6",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "christianwhocodes==1.5.2",
    "django==6.0.3",
    "django-browser-reload==1.21.0",
    "django-compressor==4.6.0",
    "django-http-compression[brotli]==1.2.0",
    "django-minify-html==1.14.0",
    "django-phonenumber-field[phonenumberslite]==8.4.0",
    "django-sass-processor==1.4.2",
    "django-watchfiles==1.4.0",
    "libsass==0.23.0",
    "pillow==12.1.1",
    "pyperclip==1.11.0",
    "python-dotenv==1.2.2",
]

[project.optional-dependencies]
vercel = ["vercel==0.5.2"]
psycopg = ["psycopg[binary,pool]==3.3.3"]

[dependency-groups]
dev = [
    "djlint>=1.36.4",
    "psycopg[binary,pool]==3.3.3",
    "ruff>=0.15.6",
    "ty>=0.0.23",
    "vercel==0.5.2",
]

[project.scripts]
tawala = "tawala.management.cli:main"
twl = "tawala.management.cli:main"

[build-system]
requires = ["uv_build>=0.10.11,<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 130
lint = { select = ["D"], ignore = ["D203", "D213"] }
format = { skip-magic-trailing-comma = true, docstring-code-format = true }

[tool.djlint]
ignore = "H016,H030,H031"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
